Final Commit

+++ server/server.js
@@ -6,6 +6,7 @@ var multer = require('multer');
 var upload = multer();
 
 const {ObjectID} = require('mongodb');
+const bcrypt=require('bcryptjs');
 
 var {mongoose} = require('./db/mongoose');
 var {Todo} = require('./models/todo');
@@ -20,9 +21,10 @@ app.use(upload.array());
 
 const port = process.env.PORT 
 
-app.post('/todos',(req,res)=>{
+app.post('/todos',authenticate,(req,res)=>{
     var newTodos = new Todo({
-        text:req.body.text
+        text:req.body.text,
+        _creator:req.user._id
     });
     
     newTodos.save().then((result)=>{
@@ -31,19 +33,19 @@ app.post('/todos',(req,res)=>{
         res.status(400).send(err);
     });
 });
-app.get('/getTodos',(req,res)=>{
-   Todo.find().then((docs)=>{
+app.get('/getTodos',authenticate,(req,res)=>{
+   Todo.find({_creator:req.user._id}).then((docs)=>{
        res.send(docs)
    }).catch((err)=>{
        res.status(400).send(err);
    }) ;
 });
 
-app.get('/todos/:id',(req,res)=>{
+app.get('/todos/:id',authenticate,(req,res)=>{
     if(!ObjectID.isValid(req.params.id)){
         return res.status(404).send();
     }
-    Todo.findById(req.params.id).then((todo)=>{
+    Todo.findOne({_id:req.params.id,_creator:req.user._id}).then((todo)=>{
         if(!todo){
             return res.status(404).send();    
         }
@@ -54,11 +56,11 @@ app.get('/todos/:id',(req,res)=>{
         res.status(404).send();
     });
 });
-app.delete('/deleteTodos/:id',(req,res)=>{
+app.delete('/deleteTodos/:id',authenticate,(req,res)=>{
     if(!ObjectID.isValid(req.params.id)){
         return res.status(400).send('Not Valid id');
     }
-    Todo.findOneAndRemove({_id:req.params.id}).then((todo)=>{
+    Todo.findOneAndRemove({_id:req.params.id,_creator:req.user._id}).then((todo)=>{
         if(!todo){
             return res.status(400).send("Not a valid id");
         }
@@ -70,7 +72,7 @@ app.delete('/deleteTodos/:id',(req,res)=>{
     });
 });
 
-app.patch('/updateTodos/:id',(req,res)=>{
+app.patch('/updateTodos/:id',authenticate,(req,res)=>{
     var id= req.params.id;
     var body = _.pick(req.body,['text','completed']);
     if(!ObjectID.isValid(id)){
@@ -85,7 +87,7 @@ app.patch('/updateTodos/:id',(req,res)=>{
         body.completedAt=null;
     }
 
-    Todo.findByIdAndUpdate(id,{$set:body},{new:true}).then((todo)=>{
+    Todo.findOneAndUpdate({_id:id,_creator:req.user._id},{$set:body},{new:true}).then((todo)=>{
         if(!todo){
             return res.status(400).send('invalid id');
         }
@@ -100,7 +102,7 @@ app.patch('/updateTodos/:id',(req,res)=>{
 app.post('/users',(req,res)=>{
     var body = _.pick(req.body,['email','password']);
     var newUser = new User(body);
-    
+    //console.log('id',newUser._id);
     newUser.save().then(()=>{
         return newUser.generateAuthToken();
     })
@@ -115,6 +117,40 @@ app.post('/users',(req,res)=>{
 
 app.get('/users/userDetails',authenticate,(req,res)=>{
     res.send(req.user);
+});
+
+
+app.post('/api/v1/todos/login',(req,res)=>{
+    var body = _.pick(req.body,['email','password']);
+
+    User.login(body.email,body.password).then((user)=>{
+        //res.status(200).send(user);
+        return user.generateAuthToken().then((token)=>{
+            res.status(200).header('x-auth',token).send(user);
+        });
+    })
+    .catch((err)=>{
+        res.status(400).send();
+    });
+
+});
+app.delete('/api/v1/todos/delete',(req,res)=>{
+    var body = _.pick(req.body,['email']);
+    User.deleteUser(body.email).then((user)=>{
+        res.status(200).send(user);
+    }).catch((err)=>{
+        //console.log(err);
+        res.status(400).send();
+    });
+});
+
+app.delete('/api/v1/todos/logout',authenticate,(req,res)=>{
+    req.user.removeToken(req.token).then(()=>{
+        res.status(200).send();
+    })
+    .catch((err)=>{
+        res.status(400).send();
+    })
 });
 
 app.listen(port,(res)=>{
